Medical Images and General bio-degradable and non bio-degradable dataset
The initial version of this dataset was imported from Kaggle and was modified to optimize the training of this model. The resulting dataset is highly effective for training any image processing or classification model related to hazardous and non-hazardous waste, as well as biodegradable and non-biodegradable waste generated in hospitals.